Avoid getting a Set-Cookie header in login response.
As determined by experimentation, the undocumented version 4 of the login API
doesn't respond with a Set-Cookie header when you don't ask for it. The docs
only go up to version 2 and my NAS reports that is has 6 versions. I found this
one by manually scanning the acceptable versions and checking their behavior.
Yay for documentation.